1. add seeds to add some coins to the user with id 1
2. modified migration file to make default value for limit float
3. preload coins when fetching user, and modified UserView correspondingly

+++ priv/repo/migrations/20180409223325_create_coins.exs
@@ -5,8 +5,8 @@ defmodule CryptoMonitor.Repo.Migrations.CreateCoins do
     create table(:coins) do
       add :code, :string
       add :user_id, references(:users, on_delete: :nothing)
-      add :limit_high, :integer
-      add :limit_low, :integer
+      add :limit_high, :float
+      add :limit_low, :float
       timestamps()
     end
